Clean uppackage configurations
