--- .dockerignore
@@ -1,10 +1,0 @@
-./.cache/
-./.devcontainer/
-.env
-./.git/
-./.github/
-./cypress/
-./cypress_php/
-./dist/
-./dist_local/
-./node_modules/
