+++ client/cypress/tsconfig.json
@@ -2,6 +2,7 @@
   "compilerOptions": {
     "target": "es5",
     "lib": ["es5", "dom"],
+    "sourceMap": true,
     "types": ["cypress", "node"]
   },
   "include": ["**/*.ts"]
