--- client/docker-compose.yml
@@ -1,19 +1,0 @@
-version: "3.9"
-
-services:
-  apache:
-    image: lyanthropos/doenet_apache:dev
-    networks:
-      - dev_net
-    depends_on:
-      - php
-    restart: always
-
-  php:
-    image: lyanthropos/doenet_php:dev
-    networks:
-      - dev_net
-    restart: always
-
-networks:
-  dev_net:
