--- client/docker/apache/doenet_dev_ssl.conf
@@ -1,179 +1,0 @@
-# This site used to use Shib, but it was disabled on 2020-06-22.
-# I'm leaving those sections in but commented in case things change.
-# -BLD
-<VirtualHost *:443>
-  ServerName dev.doenet.org
-  ServerAdmin cseithelp@umn.edu
-  Header set Access-Control-Allow-Origin "*"
-
-  DocumentRoot "/var/www/html/dev.doenet.org"
-  ## Alias declarations for resources outside the DocumentRoot
-  Alias /shibboleth-sp/main.css "/usr/share/shibboleth/main.css"
-  Alias /shibboleth-sp/logo.jpg "/usr/share/shibboleth/logo.jpg"
-
-  ## Directories, there should at least be a declaration for /var/www/html
-
-<Location "/shibboleth-sp">
-  Require all granted
-</Location>
-
-<Location "/Shibboleth.sso">
-  Require all granted
-  SetHandler shib
-</Location>
-<Directory "/var/www/html/dev.doenet.org/media/byPageId">
-  Header Set Pragma "no-cache"
-  Header Set Expires "Thu, 1 Jan 1970 00:00:00 GMT"
-  Header Set Cache-Control "max-age=0, no-store, no-cache, must-revalidate"
-  Header Unset ETag
-  FileETag None
-</Directory>
-
-  <Directory "/var/www/html/dev.doenet.org">
-    Options FollowSymLinks MultiViews
-    AllowOverride None
-    Require all granted
-    RewriteEngine on
-   # RewriteBase /
-
-   # Don't rewrite files 
-   RewriteCond %{REQUEST_FILENAME} !-f 
-   # Or anything in api
-   RewriteCond %{REQUEST_URI} !^/api [NC]
-   # or anything in media
-   RewriteCond %{REQUEST_URI} !^/media [NC]
-   # or anything in geogebra/scripts
-   RewriteCond %{REQUEST_URI} !^/geogebra/scripts [NC]
-   # or anything in /Shibboleth.sso
-   RewriteCond %{REQUEST_URI} !^/Shibboleth.sso [NC]
-   # Rewrite everything else to index.html to allow html5 state links
-   RewriteRule ^ index.html [L]
-  </Directory>
-
-  <Directory "/var/www/html/dev.doenet.org/etc/">
-    Require all denied
-  </Directory>
-
-  <Directory "/var/www/html/dev.doenet.org/admin/">
-    Options FollowSymLinks MultiViews ExecCGI
-    AllowOverride None
-#    Require valid-user
-    <FilesMatch ".+(\.cgi)$">
-        SetHandler cgi-script
-    </FilesMatch>
-#    AuthType shibboleth
-#    ShibRequestSetting requireSession 1
-#    ShibRequestSetting entityID https://idp2.shib.umn.edu/idp/shibboleth
-#    ShibRequestSetting entityID https://idp-test.shib.umn.edu/idp/shibboleth
-  </Directory>
-
-  <Directory "/var/www/html/dev.doenet.org/api/">
-    Options FollowSymLinks MultiViews ExecCGI
-    AllowOverride None
-#    Require valid-user
-    <FilesMatch ".+(\.cgi)$">
-        SetHandler cgi-script
-    </FilesMatch>
-#    AuthType shibboleth
-#    ShibRequestSetting requireSession 1
-#    ShibRequestSetting entityID https://idp2.shib.umn.edu/idp/shibboleth
-#    ShibRequestSetting entityID https://idp-test.shib.umn.edu/idp/shibboleth
-  </Directory>
-
-  <Directory "/var/www/html/dev.doenet.org/editor/">
-    Options FollowSymLinks MultiViews ExecCGI
-    AllowOverride None
-    Require valid-user
-    <FilesMatch ".+(\.cgi)$">
-        SetHandler cgi-script
-    </FilesMatch>
-    AuthType shibboleth
-    ShibRequestSetting requireSession 1
-#    ShibRequestSetting entityID https://idp2.shib.umn.edu/idp/shibboleth
-#    ShibRequestSetting entityID https://idp-test.shib.umn.edu/idp/shibboleth
-  </Directory>
-
-  <Directory "/var/www/html/dev.doenet.org/course/">
-    Options FollowSymLinks MultiViews ExecCGI
-    AllowOverride None
-#    Require valid-user
-    <FilesMatch ".+(\.cgi)$">
-        SetHandler cgi-script
-    </FilesMatch>
-#    AuthType shibboleth
-#    ShibRequestSetting requireSession 1
-#    ShibRequestSetting entityID https://idp2.shib.umn.edu/idp/shibboleth
-#    ShibRequestSetting entityID https://idp-test.shib.umn.edu/idp/shibboleth
-  </Directory>
-
-  <Directory "/var/www/html/dev.doenet.org/chooser/">
-    Options FollowSymLinks MultiViews ExecCGI
-    AllowOverride None
-  #  Require valid-user
-    <FilesMatch ".+(\.cgi)$">
-        SetHandler cgi-script
-    </FilesMatch>
-  #  AuthType shibboleth
-  #  ShibRequestSetting requireSession 1
-#    ShibRequestSetting entityID https://idp2.shib.umn.edu/idp/shibboleth
-  #  ShibRequestSetting entityID https://idp-test.shib.umn.edu/idp/shibboleth
-  </Directory>
-
-  <Directory "/var/www/html/dev.doenet.org/viewer/">
-    Options FollowSymLinks MultiViews ExecCGI
-    AllowOverride None
-  #  Require valid-user
-    <FilesMatch ".+(\.cgi)$">
-        SetHandler cgi-script
-    </FilesMatch>
-  #  AuthType shibboleth
-  #  ShibRequestSetting requireSession 1
-#    ShibRequestSetting entityID https://idp2.shib.umn.edu/idp/shibboleth
-  #  ShibRequestSetting entityID https://idp-test.shib.umn.edu/idp/shibboleth
-  </Directory>
-
-  <Directory "/var/www/html/dev.doenet.org/live/">
-    Options FollowSymLinks MultiViews ExecCGI
-    AllowOverride None
-#    Require valid-user
-    <FilesMatch ".+(\.cgi)$">
-        SetHandler cgi-script
-    </FilesMatch>
-#    AuthType shibboleth
-#    ShibRequestSetting requireSession 1
-#    ShibRequestSetting entityID https://idp2.shib.umn.edu/idp/shibboleth
-#    ShibRequestSetting entityID https://idp-test.shib.umn.edu/idp/shibboleth
-  </Directory>
-
-  <Directory "/var/www/html/dev.doenet.org/open_api/">
-    Options -Indexes
-  </Directory>
-
-  #<Directory "/var/www/html/dev.doenet.org/">
-  <Directory "/var/www/html/dev.doenet.org/umn">
-    Require valid-user
-    AuthType shibboleth
-    ShibRequestSetting requireSession 1
-  </Directory> 
-
-  <Directory "/var/www/html/dev.doenet.org/api/umn">
-    Require valid-user
-    AuthType shibboleth
-    ShibRequestSetting requireSession 1
-  </Directory> 
-
-
-  ## Logging
-  ErrorLog "/var/log/httpd/dev.doenet.org_ssl_error.log"
-  ServerSignature Off
-  CustomLog "/var/log/httpd/dev.doenet.org_ssl_access.log" combined
-
-  ## Server aliases
-  ServerAlias cse-doenet-dev-dev.oit.umn.edu
-
-  ## SSL directives
-  SSLEngine on
-  SSLCertificateFile      "/etc/httpd/ssl/wildcard.doenet.org.crt"
-  SSLCertificateKeyFile   "/etc/httpd/ssl/wildcard.doenet.org.key"
-  SSLCACertificateFile    "/etc/httpd/ssl/fullchain.pem"
-</VirtualHost>
