--- client/docker/docker-compose-production.yml
@@ -1,27 +1,0 @@
-version: "3.9"
-
-services:
-  apache:
-    image: lyanthropos/doenet_test:apache-dev
-    ports:
-      - 80:80
-    volumes:
-      - ./dist:/usr/local/apache2/htdocs/
-      # Likely need to mount an env.ini file for key access
-    networks:
-      - dev_net
-    depends_on:
-      - php
-    restart: always
-
-  php:
-    image: lyanthropos/doenet_test:php-dev
-    networks:
-      - dev_net
-    restart: always
-
-volumes:
-  vendor:
-
-networks:
-  dev_net:
