--- client/docker/php/composer.json
@@ -1,5 +1,0 @@
-{
-  "require": {
-    "firebase/php-jwt": "5.2.*"
-  }
-}
