--- client/docker/php/env.ini
@@ -1,7 +1,0 @@
-dbhost=mysql
-database=doenet_local
-username=root
-password=helloworld
-mode=development
-key=super_secret_key
-web3storagetoken=update_this_to_your_token
