+++ client/eslint.config.mjs
@@ -1,17 +1,29 @@
 import globals from "globals";
 import pluginJs from "@eslint/js";
 import tseslint from "typescript-eslint";
+import unusedImports from "eslint-plugin-unused-imports";
 
 export default [
   { files: ["**.{js,mjs,cjs,ts,jsx,tsx}"] },
-  { ignores: ["dist/", "dist_local"] },
+  { ignores: ["dist/"] },
   { languageOptions: { globals: { ...globals.node } } },
+  {
+    plugins: {
+      "unused-imports": unusedImports,
+    },
+  },
   pluginJs.configs.recommended,
   ...tseslint.configs.recommended,
   {
     rules: {
-      "@typescript-eslint/no-unused-vars": [
-        "error",
+      // For now, we will allow explicit anys, but
+      // we should re-enable this rule in the future.
+      "@typescript-eslint/no-explicit-any": "off",
+
+      "@typescript-eslint/no-unused-vars": "off",
+      "unused-imports/no-unused-imports": "error",
+      "unused-imports/no-unused-vars": [
+        "warn",
         {
           args: "all",
           argsIgnorePattern: "^_",
@@ -22,9 +34,6 @@ export default [
           ignoreRestSiblings: true,
         },
       ],
-      "@typescript-eslint/no-explicit-any": "off",
-      "@typescript-eslint/ban-ts-comment": "off",
-      "@typescript-eslint/no-namespace": "off",
     },
   },
 ];
