+++ client/package.json
@@ -6,25 +6,12 @@
     "dev": "vite",
     "preview": "view preview",
     "build": "vite build",
-    "dc": "docker compose -f ./docker-compose.yml -f ./.devcontainer/docker-compose.extend.yml",
-    "artifactory:build": "docker buildx build --platform linux/arm64,linux/amd64 -t cse-docker-doenet.artifactory.umn.edu/doenet_$npm_config_service:$npm_package_version -f docker/$npm_config_service/Dockerfile -t cse-docker-doenet.artifactory.umn.edu/doenet_$npm_config_service --provenance=false .",
-    "artifactory:push": "npm run artifactory:build -- --push",
-    "deploy:local": "npm run dc up -- -d --no-deps --build apache",
-    "deploy:development": "npm run build -- --mode development --outDir dist --minify false",
-    "deploy:production": "npm run build -- --mode production --outDir dist",
-    "deploy:prod": "npm run deploy:production",
-    "deploy:dev": "npm run deploy:development",
-    "start:ci": "docker compose up -d",
-    "start:db": "docker compose exec mysql mysql -u root -phelloworld -e 'CREATE DATABASE IF NOT EXISTS doenet_local CHARACTER SET=utf8 COLLATE=utf8_unicode_ci;'",
     "test": "cypress open",
     "test:all": "cypress run -b 'chrome' --config video=false --headless",
     "test:app": "cypress run -b 'electron' --config video=false,excludeSpecPattern=cypress/e2e/DoenetML/** --headless",
     "test:parallel": "cypress-parallel -s test:all -t 4 -d cypress/e2e/DoenetML",
-    "publish:db": "docker compose exec -T -w /var/lib/mysql mysql mysqldump -u root -phelloworld --databases --add-drop-database --add-drop-table doenet_local > ./docker/volumes/db_init/db_template.sql",
-    "reset:db": "docker compose exec -T mysql mysql -u root -phelloworld doenet_local < ./docker/mysql/db_template.sql",
-    "reset:volumes": "docker volume rm doenet_node_modules doenet_vendor",
     "format": "prettier --write \"src/**/*.{js,jsx,ts,tsx,json,css}\" \"cypress/**/*.{js,jsx,ts,tsx,json,css}\" --end-of-line auto",
-    "lint": "eslint"
+    "lint": "eslint --fix"
   },
   "dependencies": {
     "@chakra-ui/icons": "^2.2.4",
@@ -32,8 +19,6 @@
     "@doenet/assignment-viewer": "^0.1.0-alpha6",
     "@doenet/doenetml-iframe": "^0.7.0-alpha39",
     "@fontsource/jost": "^5.2.5",
-    "@fortawesome/free-solid-svg-icons": "^6.7.2",
-    "@fortawesome/react-fontawesome": "^0.2.2",
     "axios": "^1.8.4",
     "better-react-mathjax": "^2.1.0",
     "browser-image-resizer": "^2.4.1",
@@ -54,9 +39,8 @@
     "react-router": "^7.4.1",
     "react-select": "^5.10.1",
     "recharts": "^2.15.1",
-    "styled-components": "5.3.11",
-    "swiper": "^9.4.1",
-    "typescript-eslint": "^8.28.0"
+    "styled-components": "6.1.18",
+    "swiper": "^9.4.1"
   },
   "devDependencies": {
     "@babel/eslint-parser": "^7.27.0",
@@ -81,12 +65,14 @@
     "eslint-plugin-jsx-a11y": "^6.10.2",
     "eslint-plugin-react": "^7.37.4",
     "eslint-plugin-react-hooks": "^5.2.0",
+    "eslint-plugin-unused-imports": "^4.1.4",
     "http-proxy": "^1.18.1",
     "mysql2": "^3.14.0",
     "prettier": "^3.5.3",
     "rollup": "^4.39.0",
     "rollup-plugin-polyfill-node": "^0.13.0",
     "typescript": "^5.8.2",
+    "typescript-eslint": "^8.33.0",
     "vite": "^6.2.4",
     "vite-plugin-mpa": "^1.2.0",
     "vite-plugin-static-copy": "^2.3.0"
@@ -100,88 +86,5 @@
     "cypress-plugin-tab": "^1.0.5",
     "cypress-wait-until": "^3.0.2",
     "eslint-plugin-cypress": "^4.2.0"
-  },
-  "eslintConfig": {
-    "extends": [
-      "eslint:recommended",
-      "plugin:import/errors",
-      "plugin:react/recommended",
-      "plugin:jsx-a11y/recommended",
-      "plugin:cypress/recommended",
-      "prettier"
-    ],
-    "rules": {
-      "linebreak-style": [
-        "error",
-        "unix"
-      ],
-      "react/prop-types": "off",
-      "no-console": "warn",
-      "react-hooks/rules-of-hooks": "error",
-      "react-hooks/exhaustive-deps": [
-        "warn"
-      ]
-    },
-    "plugins": [
-      "react",
-      "import",
-      "jsx-a11y",
-      "react-hooks",
-      "cypress"
-    ],
-    "overrides": [
-      {
-        "files": [
-          "./src/Core/**/*.js",
-          "./src/Viewer/**/*.{jsx,js}"
-        ],
-        "parser": "@babel/eslint-parser",
-        "parserOptions": {
-          "sourceType": "module",
-          "requireConfigFile": false,
-          "babelOptions": {
-            "plugins": [
-              "@babel/plugin-syntax-class-properties"
-            ]
-          }
-        }
-      }
-    ],
-    "parserOptions": {
-      "ecmaVersion": 11,
-      "sourceType": "module",
-      "ecmaFeatures": {
-        "jsx": true
-      }
-    },
-    "env": {
-      "browser": true,
-      "es2020": true,
-      "node": true
-    },
-    "settings": {
-      "import/core-modules": [
-        "@ToolRoot",
-        "@Tool",
-        "solid-svg",
-        "react-spring"
-      ],
-      "import/resolver": {
-        "node": {
-          "paths": [
-            "src"
-          ],
-          "extensions": [
-            ".js",
-            ".jsx",
-            ".ts",
-            ".tsx"
-          ]
-        }
-      },
-      "react": {
-        "version": "detect"
-      }
-    }
   }
 }
