+++ client/src/_utils/responses.tsx
@@ -1,8 +1,9 @@
 import React from "react";
+//@ts-expect-error Math expressions is not typed
 import me from "math-expressions";
 import { MathJax } from "better-react-mathjax";
 
-const nanInfinityReviver = function (_key, value) {
+const nanInfinityReviver = function (_key: any, value: any) {
   if (value && value.objectType === "special-numeric") {
     if (value.stringValue === "NaN") {
       return NaN;
@@ -19,17 +20,14 @@ const nanInfinityReviver = function (_key, value) {
 export function parseAndFormatResponse(response: string): React.JSX.Element {
   const parsedResp = JSON.parse(response, nanInfinityReviver);
 
-  return parsedResp.response.map((v, i) => {
+  return parsedResp.response.map((v: any, i: any) => {
     const componentType = parsedResp.componentTypes[i];
     if (["math", "point", "matrix", "vector"].includes(componentType)) {
       const expr = me.fromAst(v).round_numbers_to_precision_plus_decimals(6, 2);
       return (
         <div key={i}>
           <MathJax hideUntilTypeset={"first"} inline dynamic key={i}>
-            {
-              //@ts-ignore
-              "\\(" + expr.toLatex() + "\\)"
-            }
+            {"\\(" + expr.toLatex() + "\\)"}
           </MathJax>
         </div>
       );
