+++ server/eslint.config.mjs
@@ -1,17 +1,25 @@
 import globals from "globals";
 import pluginJs from "@eslint/js";
 import tseslint from "typescript-eslint";
+import unusedImports from "eslint-plugin-unused-imports";
 
 export default [
   { files: ["**.{js,mjs,cjs,ts,jsx,tsx}"] },
   { ignores: ["dist/"] },
   { languageOptions: { globals: { ...globals.node } } },
+  {
+    plugins: {
+      "unused-imports": unusedImports,
+    },
+  },
   pluginJs.configs.recommended,
   ...tseslint.configs.recommended,
   {
     rules: {
-      "@typescript-eslint/no-unused-vars": [
-        "error",
+      "@typescript-eslint/no-unused-vars": "off",
+      "unused-imports/no-unused-imports": "error",
+      "unused-imports/no-unused-vars": [
+        "warn",
         {
           args: "all",
           argsIgnorePattern: "^_",
