+++ server/package.json
@@ -9,7 +9,7 @@
     "build": "npx tsc && cp src/signin_email.html dist/src/",
     "start": "node dist/src/index.js",
     "format": "npx prettier . --write --end-of-line auto",
-    "lint": "npx eslint"
+    "lint": "npx eslint --fix"
   },
   "keywords": [],
   "author": "",
@@ -47,6 +47,7 @@
     "@types/passport-google-oauth20": "^2.0.16",
     "cookie-session": "^2.1.0",
     "eslint": "^9.23.0",
+    "eslint-plugin-unused-imports": "^4.1.4",
     "globals": "^16.0.0",
     "nodemon": "^3.1.9",
     "prettier": "^3.5.3",
