+++ server/src/types/passport-magic-link.d.ts
@@ -63,11 +63,11 @@ declare module "passport" {
   }
 
   interface Authenticator<
-    // eslint-disable-next-line @typescript-eslint/no-unused-vars
+    // eslint-disable-next-line unused-imports/no-unused-vars
     InitializeRet = express.Handler,
     AuthenticateRet = any,
     AuthorizeRet = AuthenticateRet,
-    // eslint-disable-next-line @typescript-eslint/no-unused-vars
+    // eslint-disable-next-line unused-imports/no-unused-vars
     AuthorizeOptions = passport.AuthenticateOptions,
   > {
     authenticate(
